+++ crates/test-utils/src/lib.rs
@@ -346,10 +346,8 @@ impl Test {
             (Assembler::new(self.source_manager.clone()), None)
         };
 
-        let mut assembler = self
-            .add_modules
-            .iter()
-            .fold(assembler, |mut assembler, (path, source)| {
+        let mut assembler =
+            self.add_modules.iter().fold(assembler, |mut assembler, (path, source)| {
                 let module = source
                     .parse_with_options(
                         &self.source_manager,
@@ -358,8 +356,8 @@ impl Test {
                     .expect("invalid masm source code");
                 assembler.compile_and_statically_link(module).expect("failed to link module");
                 assembler
-            })
-            .with_debug_mode(self.in_debug_mode);
+            });
+        // Debug mode is now always enabled
         for library in &self.libraries {
             assembler.link_dynamic_library(library).unwrap();
         }