+++ miden-vm/src/cli/prove.rs
@@ -104,8 +104,7 @@ impl ProveCmd {
         let (program, mut host) = match ext.as_str() {
             "masp" => (get_masp_program(&self.program_file)?, host),
             "masm" => {
-                let (program, source_manager) =
-                    get_masm_program(&self.program_file, &libraries, true)?;
+                let (program, source_manager) = get_masm_program(&self.program_file, &libraries)?;
                 (program, host.with_source_manager(source_manager))
             },
             _ => return Err(Report::msg("The provided file must have a .masm or .masp extension")),