+++ miden-vm/tests/integration/operations/decorators/events.rs
@@ -85,8 +85,7 @@ fn test_debug_with_debugging() {
     end";
 
     // compile and execute program
-    let program: Program =
-        Assembler::default().with_debug_mode(true).assemble_program(source).unwrap();
+    let program: Program = Assembler::default().assemble_program(source).unwrap();
     let mut host = TestHost::default();
     miden_processor::execute(
         &program,
@@ -141,8 +140,7 @@ fn test_parsing_debug_advice_stack() {
     end";
 
     // compile and execute program
-    let program: Program =
-        Assembler::default().with_debug_mode(true).assemble_program(source).unwrap();
+    let program: Program = Assembler::default().assemble_program(source).unwrap();
     let mut host = TestHost::default();
     miden_processor::execute(
         &program,