+++ processor/src/tests/mod.rs
@@ -736,13 +736,9 @@ fn test_diagnostic_no_mast_forest_with_procedure() {
         end
     ";
 
-    let library = Assembler::new(source_manager.clone())
-        .with_debug_mode(true)
-        .assemble_library([lib_module])
-        .unwrap();
+    let library = Assembler::new(source_manager.clone()).assemble_library([lib_module]).unwrap();
 
     let program = Assembler::new(source_manager.clone())
-        .with_debug_mode(true)
         .with_dynamic_library(&library)
         .unwrap()
         .assemble_program(program_source)
@@ -962,13 +958,10 @@ fn test_diagnostic_syscall_target_not_in_kernel() {
         end
     ";
 
-    let kernel_library = Assembler::new(source_manager.clone())
-        .with_debug_mode(true)
-        .assemble_kernel(kernel_source)
-        .unwrap();
+    let kernel_library =
+        Assembler::new(source_manager.clone()).assemble_kernel(kernel_source).unwrap();
 
     let program = Assembler::with_kernel(source_manager.clone(), kernel_library)
-        .with_debug_mode(true)
         .assemble_program(program_source)
         .unwrap();